refactor: userDetails 클래스 user에서 상속받기
